ARM: shmobile: sh73a0 dtsi: Add Cortex-A9 TWD node
Add a node for the Private Timer and Watchdog, as found in the Cortex-A9
MPCore.

Without this, there's no clocksource available during early kernel
initialization, before cmt1 is initialized, leading to a lock-up if
CONFIG_CPU_IDLE=y.

Signed-off-by: Geert Uytterhoeven <geert+renesas@glider.be>
---
As the TWD driver does not support DT without CCF, this should NOT be
applied to a branch that still has kzm9g-reference support.
Hence the first choice "dt-for-v3.20" is NOT OK.
"sh73a0-multiplatform-for-v3.20" is OK.

v2:
  - The clock is twd_clk