clk: shmobile: Fix r8a7795 MSSR support
This is an incremental fix for the series posted yesterday:
[PATCH v8 00/05] Renesas R-Car Gen3 CPG support V8

As pointed out by Geert, I missed one r8a7791 instance
when doing manual search-and-replace. This fix will in
the future be folded into patch 3/5 and 4/5, but for
now this incremental patch will have to do.

Thanks to Geert for keeping his eyes open!

Signed-off-by: Magnus Damm <damm+renesas@opensource.se>
Signed-off-by: Geert Uytterhoeven <geert+renesas@glider.be>